requirements.txt
about-time==4.2.1
aiohttp==3.8.4
aiosignal==1.3.1
alive-progress==3.1.1
astroid==2.5
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
bibtexparser==1.4.0
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.1.0
colorama==0.4.6
crossref-commons==0.0.7
frozenlist==1.3.3
future==0.18.3
grapheme==0.6.0
HTMLParser==0.0.2
idna==2.10
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.6.1
multidict==6.0.4
numpy==1.24.3
openai==0.27.4
pandas==2.0.1
pyChainedProxy==1.2
pylint==2.7.1
PyPaperBot==1.2.2
pyparsing==3.0.9
pypdf==3.8.0
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
ratelimit==2.2.1
regex==2023.6.3
requests==2.28.2
six==1.16.0
soupsieve==2.4.1
tabulate==0.9.0
tiktoken==0.4.0
toml==0.10.2
tzdata==2023.3
urllib3==1.26.15
wrapt==1.12.1
yarl==1.8.2
